refactor: Client 타입과 Bank 타입의 파일분리 #3
