+++ BankManagerConsoleApp/BankManagerConsoleApp/Client.swift
@@ -1,0 +1,17 @@
+//
+//  Client.swift
+//  BankManagerConsoleApp
+//
+//  Created by Sunwoo on 2021/12/27.
+//
+
+import Foundation
+
+struct Ticket {
+    let number: Int
+}
+
+struct Client {
+    let orderTicket: Ticket
+    let taskDuration = 0.7
+}
